add auth middleware

+++ graphql/resolvers/booking.js
@@ -3,9 +3,14 @@ const Booking = require('../../models/booking');
 const { transformBooking, transformEvent } = require('./merge');
 
 module.exports = {
-  bookings: async () => {
+  bookings: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated !');
+    }
+
     try {
       const bookings = await Booking.find();
+
       return bookings.map((booking) => {
         return transformBooking(booking);
       });
@@ -13,20 +18,33 @@ module.exports = {
       throw err;
     }
   },
-  bookEvent: async (args) => {
+
+  bookEvent: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated !');
+    }
+
     const fetchedEvent = await Event.findOne({ _id: args.eventId });
     const booking = new Booking({
       user: '5c7264363e26c827003ac771',
       event: fetchedEvent
     });
+
     const result = await booking.save();
+
     return transformBooking(result);
   },
-  cancelBooking: async (args) => {
+
+  cancelBooking: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated !');
+    }
+
     try {
       const booking = await Booking.findById(args.bookingId).populate('event');
       const event = transformEvent(booking.event);
       await Booking.deleteOne({ _id: args.bookingId });
+
       return event;
     } catch (err) {
       throw err;
